For alpine, non-amd64 install regular knock package

+++ build/alpine/install-packages.sh
@@ -3,6 +3,13 @@
 set -e
 set -o pipefail
 
+# Patched knockd is only available for amd64 on Alpine, so install the package for other architectures
+knockPkg=""
+if [ "$TARGETARCH" != amd64 ]; then
+  knockPkg="knock"
+  echo "Installing knock package for $TARGETARCH architecture"
+fi
+
 # Install necessary packages
 # shellcheck disable=SC2086
 apk add --no-cache -U \
@@ -36,13 +43,18 @@ apk add --no-cache -U \
     numactl \
     jattach \
     gcompat \
+    "$knockPkg" \
     ${EXTRA_ALPINE_PACKAGES}
 
-# Download and install patched knockd
-curl -fsSL -o /tmp/knock.tar.gz "https://github.com/${KNOCKD_REPO_ORG}/releases/download/${KNOCKD_VERSION}/knock-${KNOCKD_VERSION}-alpine-amd64.tar.gz"
-tar -xf /tmp/knock.tar.gz -C /usr/local/ && rm /tmp/knock.tar.gz
-ln -s /usr/local/sbin/knockd /usr/sbin/knockd
-setcap cap_net_raw=ep /usr/local/sbin/knockd
+if ! [ "$knockPkg" ]; then
+  # Download and install patched knockd
+  knockdUrl="https://github.com/${KNOCKD_REPO_ORG}/releases/download/${KNOCKD_VERSION}/knock-${KNOCKD_VERSION}-alpine-amd64.tar.gz"
+  echo "Downloading knockd from $knockdUrl"
+  curl -fsSL -o /tmp/knock.tar.gz "$knockdUrl"
+  tar -xf /tmp/knock.tar.gz -C /usr/local/ && rm /tmp/knock.tar.gz
+  ln -s /usr/local/sbin/knockd /usr/sbin/knockd
+  setcap cap_net_raw=ep /usr/local/sbin/knockd
+fi
 
 # Set Git credentials globally
 cat <<EOF >> /etc/gitconfig
